% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_simulation.R
\name{abs_diff}
\alias{abs_diff}
\title{Calculate Absolute difference in true-median beta}
\usage{
abs_diff(model, par)
}
\arguments{
\item{model}{stapreg object from rstap package}

\item{par}{parameter named vector parameter
with true parameter value as elements}
}
\description{
Calculate Absolute difference in true-median beta
}
