% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_simulation.R
\name{calculate_CG_stat}
\alias{calculate_CG_stat}
\title{calculate C&G validation statistic}
\usage{
calculate_CG_stat(model, par, mer = FALSE)
}
\arguments{
\item{model}{stapreg object from rstap package}

\item{par}{named vector}
}
\description{
calculate C&G validation statistic
}
