% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_simulation.R
\name{calculate_RMSE_median}
\alias{calculate_RMSE_median}
\title{calculate Median Root Mean Square Error}
\usage{
calculate_RMSE_median(model)
}
\arguments{
\item{model}{stapreg object from rstap package}
}
\description{
calculate Median Root Mean Square Error
}
