% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_simulation.R
\name{check_coverage}
\alias{check_coverage}
\title{check coverage}
\usage{
check_coverage(model, true_par, prob = 0.9, nonzero = T)
}
\arguments{
\item{model}{stapreg object from rstap package}

\item{true_par}{a named vector with the true parameter values as elements}

\item{prob}{probability value in (0,1)}

\item{nonzero}{logical indicating zero should not be in the posterior interval}
}
\description{
check coverage
}
