% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{create_table_one}
\alias{create_table_one}
\title{create_table_one in AoAS paper}
\usage{
create_table_one(num_sims = 5, n = 300, num_MESA_subj = 100,
  d = 20, alpha = 25, theta = 0.5, delta = -2.2, beta = 0.75,
  beta_bar = 0.85, sigma = 1, alpha_prior = rstap::normal(location =
  25, autoscale = F), beta_prior = rstap::normal(),
  theta_prior = rstap::log_normal(location = 1, scale = 1),
  delta_prior = rstap::normal(), iter = 2000, warmup = 1000,
  cores = 1, chains = 1, seed = NULL, file = NULL)
}
\arguments{
\item{num_sims}{number of simulations to run}

\item{n}{number of subjects to simulate}

\item{d}{number of distances to simulate}

\item{alpha}{intercept for generating outcome}

\item{theta}{true spatial scale under which datasets}

\item{delta}{simulated binary covariate regression effect}

\item{beta}{SAP effect}

\item{alpha_prior}{prior to be placed on intercept in model}

\item{beta_prior}{prior to be placed on SAP effect}

\item{theta_prior}{prior to be placed on spatial scale}

\item{delta_prior}{prior to be placed on simulated binary covariate effect}
}
\value{
list of two table components table1_top -
which includes the coverage and diagnostic statistics broken down by parameter and
table1_bottom which has the RMSE for the entire model
}
\description{
create_table_one in AoAS paper
}
