% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{create_table_two}
\alias{create_table_two}
\title{create_table_two}
\usage{
create_table_two(num_sims = 5, num_subj = 100, num_dists = 30,
  alpha = 25, theta = 0.5, shape = 5, delta = -2.2, beta = 0.75,
  iter = 2000, warmup = 1000, chains = 1, cores = 1, file = NULL)
}
\arguments{
\item{num_sims}{}
}
\description{
create_table_two
}
