% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DLM_dataset.R
\name{generate_dlm_dataset}
\alias{generate_dlm_dataset}
\title{generates Distributed Lag Model or polynomial function effect}
\usage{
generate_dlm_dataset(seed, num_subj = 100L, Lambda_xy = function(x, y)
  {     20 - x^2 + y^2 }, max_dist = 5L, bef_xlim = c(0, 1),
  bef_ylim = c(0, 1), subj_xlim = c(0, 1), subj_ylim = c(0, 1),
  alpha = 25, delta = -2.3, beta = 0.1, theta = 0.5, sigma = 1,
  W = function(x) exp(-x))
}
\arguments{
\item{seed}{an integer for set.seed()}

\item{num_subj}{number of subjects to simulate}

\item{max_dist}{upper bound on distances to include}

\item{bef_xlim}{x coordinate bounds from which to simulate befs}

\item{bef_ylim}{y coordinate bounds from which to simulate bef positions}

\item{subj_xlim}{x coordinate bounds from which to simulate subject positions}

\item{subj_ylim}{y coordinate bounds from which to simulate subject positions}

\item{alpha}{true intercept value}

\item{delta}{confounder effect}

\item{beta}{true bef effect}

\item{theta}{true spatial scale}

\item{W}{weight function}

\item{num_dists}{number of distances/subject to simulate}

\item{inten}{intensity function for nhpp as a function of spatial coords (x,y)}
}
\description{
generates Distributed Lag Model or polynomial function effect
}
