% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_hpp_dataset.R
\name{generate_hpp_dataset}
\alias{generate_hpp_dataset}
\title{generates homogenous poisson process data}
\usage{
generate_hpp_dataset(seed, num_subj = 100L, num_dists = 30L,
  max_dist = 5L, bef_xlim = c(-1, 1), bef_ylim = c(-1, 1),
  subj_xlim = c(-1, 1), subj_ylim = c(-1, 1), alpha = 25,
  delta = -2.3, beta = 0.1, theta = 0.5, shape = NULL, sigma = 1,
  K = function(x) exp(-x))
}
\arguments{
\item{seed}{an integer for set.seed()}

\item{num_subj}{number of subjects to simulate}

\item{num_dists}{number of distances/subject to simulate}

\item{max_dist}{upper bound on distances to include}

\item{bef_xlim}{x coordinate bounds from which to simulate befs}

\item{bef_ylim}{y coordinate bounds from which to simulate bef positions}

\item{subj_xlim}{x coordinate bounds from which to simulate subject positions}

\item{subj_ylim}{y coordinate bounds from which to simulate subject positions}

\item{alpha}{true intercept value}

\item{delta}{confounder effect}

\item{beta}{true bef effect}

\item{theta}{true spatial scale}

\item{K}{exposure function}
}
\description{
generates homogenous poisson process data
}
