% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nhpp_dataset.R
\name{generate_nhpp_dataset}
\alias{generate_nhpp_dataset}
\title{generates nonhomogenous poisson process data}
\usage{
generate_nhpp_dataset(seed = NULL, num_subj = 5L, num_dists = 2L,
  Lambda_xy = function(x, y) {     (3 - x^2 - y^2) }, max_dist = 5L,
  bef_xlim = c(-1, 1), bef_ylim = c(-1, 1), subj_xlim = c(-1, 1),
  subj_ylim = c(-1, 1), alpha = 25, delta = -2.3, beta = 0.1,
  theta = 0.5, sigma = 1, K = function(x) exp(-x))
}
\arguments{
\item{seed}{an integer for set.seed()}

\item{num_subj}{number of subjects position processes to simulate}

\item{num_dists}{number of processes to simulate}

\item{max_dist}{upper bound on distances to include}

\item{bef_xlim}{x coordinate bounds from which to simulate befs}

\item{bef_ylim}{y coordinate bounds from which to simulate bef positions}

\item{subj_xlim}{x coordinate bounds from which to simulate subject positions}

\item{subj_ylim}{y coordinate bounds from which to simulate subject positions}

\item{alpha}{true intercept value}

\item{delta}{confounder effect}

\item{beta}{true bef effect}

\item{theta}{true spatial scale}

\item{K}{exposure function}

\item{inten}{intensity function for nhpp as a function of spatial coords (x,y)}
}
\description{
generates nonhomogenous poisson process data
}
