% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_nhpp_table_one.R
\name{generate_nhpp_table_one}
\alias{generate_nhpp_table_one}
\title{generates nonhomogenous poisson process data without residual error for table one}
\usage{
generate_nhpp_table_one(seed = NULL, num_subj_realizations = 10L,
  num_dist_realizations = 4L, Lambda_xy = function(x, y) {     (3 - x^2
  - y^2) }, bef_xlim = c(-1, 1), bef_ylim = c(-1, 1),
  subj_xlim = c(-1, 1), subj_ylim = c(-1, 1), alpha = 23,
  delta = -2.2, beta = 0.75, theta = 0.5, sigma = 1,
  K = function(x) exp(-x))
}
\arguments{
\item{seed}{an integer for set.seed()}

\item{num_subj_realizations}{number of subjects position processes to simulate}

\item{num_dist_realizations}{number of bef processes to simulate}

\item{bef_xlim}{x coordinate bounds from which to simulate befs}

\item{bef_ylim}{y coordinate bounds from which to simulate bef positions}

\item{subj_xlim}{x coordinate bounds from which to simulate subject positions}

\item{subj_ylim}{y coordinate bounds from which to simulate subject positions}

\item{alpha}{true intercept value}

\item{delta}{confounder effect}

\item{beta}{true bef effect}

\item{theta}{true spatial scale}

\item{K}{exposure function}

\item{inten}{intensity function for nhpp as a function of spatial coords (x,y)}
}
\description{
generates nonhomogenous poisson process data without residual error for table one
}
