% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_simulation.R
\name{interval_length}
\alias{interval_length}
\title{calculate interval length}
\usage{
interval_length(model, par, prob = 0.95)
}
\arguments{
\item{model}{stapreg object from rstap package}

\item{par}{parameter names as vector of strings}

\item{prob}{scalar numeric in (0,1)}
}
\description{
calculate interval length
}
