% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table_one.R
\name{create_table_one}
\alias{create_table_one}
\title{create_table_one in AoAS paper}
\usage{
create_table_one(num_sims = 5, num_subj = 50L, num_mesa_subj = 50L,
  num_dists = 30, alpha = 23, theta = 0.5, delta = -2.2,
  beta = 0.75, beta_bar = 0.85, alpha_prior = rstap::normal(location
  = 25, scale = 4, autoscale = F), beta_prior = rstap::normal(location =
  0, scale = 3, autoscale = F), theta_prior = rstap::log_normal(location
  = 0, scale = 1), delta_prior = rstap::normal(location = 0, scale = 3,
  autoscale = F), iter = 4000, warmup = 2000, chains = 1,
  cores = 1, file = NULL)
}
\arguments{
\item{num_sims}{number of simulations to run}

\item{num_subj}{number of subjects to simulate}

\item{num_mesa_subj}{number of subjects to sample from MESA data for MESA analysis}

\item{num_dists}{number of distances to simulate}

\item{alpha}{intercept for generating outcome}

\item{theta}{true spatial scale under which datasets}

\item{delta}{simulated binary covariate regression effect}

\item{beta}{SAP effect}

\item{alpha_prior}{prior to be placed on intercept in model, must be an rstap:: namespace object}

\item{beta_prior}{prior to be placed on SAP effect}

\item{theta_prior}{prior to be placed on spatial scale}

\item{delta_prior}{prior to be placed on simulated binary covariate effect}

\item{iter}{number of iterations for which to run the stap_glm or stapdnd_glmer sampler}

\item{warmup}{number of iterations to warmup the sampler}

\item{chains}{number of independent MCMC chains to draw}

\item{cores}{number of cores with which to run chains in parallel}

\item{file}{path to file to save tables to in .tex format}
}
\value{
list of four table components table1_top -
which includes the coverage and diagnostic statistics broken down by parameter and
table1_bottom which has the RMSE for the entire model.
The two remaining "raw" table components contain the pre-aggregation data-frames
}
\description{
create_table_one in AoAS paper
}
