% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{create_table_two}
\alias{create_table_two}
\title{create_table_two}
\usage{
create_table_two(num_sims = 5, num_subj = 300, num_dists = 20,
  alpha = 25, theta = 0.5, shape = 5, delta = -2.2, beta = 0.75,
  cores = 2)
}
\arguments{
\item{num_sims}{}
}
\description{
create_table_two
}
