% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_DLM_dataset.R
\name{generate_dlm_dataset}
\alias{generate_dlm_dataset}
\title{generates Distributed Lag Model or polynomial function effect}
\usage{
generate_dlm_dataset(seed, num_subj = 100L, num_dists = 50L,
  max_dist = 10L, bef_xlim = c(0L, 1L), bef_ylim = c(0L, 1L),
  subj_xlim = c(0L, 1L), subj_ylim = c(0L, 1L), alpha = 25L,
  delta = -2.3, beta = 0.1, kern = 20, theta = 40,
  residual_error = 1)
}
\arguments{
\item{seed}{an integer for set.seed()}

\item{num_subj}{number of subjects to simulate}

\item{num_dists}{number of distances/subject to simulate}

\item{max_dist}{upper bound on distances to include}

\item{bef_xlim}{x coordinate bounds from which to simulate befs}

\item{bef_ylim}{y coordinate bounds from which to simulate bef positions}

\item{subj_xlim}{x coordinate bounds from which to simulate subject positions}

\item{subj_ylim}{y coordinate bounds from which to simulate subject positions}

\item{alpha}{true intercept value}

\item{delta}{confounder effect}

\item{beta}{true bef effect}

\item{theta}{true spatial scale}

\item{inten}{intensity function for nhpp as a function of spatial coords (x,y)}

\item{W}{weight function}
}
\description{
generates Distributed Lag Model or polynomial function effect
}
