% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dlm_dataset.R
\name{generate_dlm_dataset}
\alias{generate_dlm_dataset}
\title{generates Distributed Lag Model or polynomial function effect}
\usage{
generate_dlm_dataset(seed = NULL, num_subj = 100, num_dists = 30,
  bef_xlim = c(-1, 1), bef_ylim = c(-1, 1), subj_xlim = c(-1, 1),
  subj_ylim = c(-1, 1), alpha = 25, delta = -2.3, beta = 0.1,
  sigma = 1, K = function(x) {     (x <= 1.2) * 1 })
}
\arguments{
\item{seed}{an integer for set.seed()}

\item{num_subj}{number of subjects to simulate}

\item{num_dists}{number of distances/subject to simulate}

\item{alpha}{true intercept value}

\item{delta}{confounder effect}

\item{beta}{true bef effect}

\item{theta}{true spatial scale}

\item{W}{weight function}
}
\description{
generates Distributed Lag Model or polynomial function effect
}
