% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mesa_dataset.R
\name{generate_mesa_dataset}
\alias{generate_mesa_dataset}
\title{generates stap dataset using MESA HFS-cohort distances}
\usage{
generate_mesa_dataset(seed = NULL, num_subj = 100, prop_dist = 0.95,
  max_dist = 10, alpha = 23, delta = -2.2, beta = 0.1,
  beta_bar = 0.5, theta = 0.7, sigma = 1, K = function(x) exp(-x))
}
\arguments{
\item{seed}{an integer for set.seed()}

\item{num_subj}{number of subjects to simulate}

\item{prop_dist}{proportion of MESA distances to use -
set to 1 for all MESA distances, or smaller to speed model fitting time}

\item{max_dist}{upper bound on distances to include}

\item{alpha}{true intercept value}

\item{delta}{confounder effect}

\item{beta}{true bef effect}

\item{theta}{true spatial scale}

\item{K}{exposure function}
}
\description{
generates stap dataset using MESA HFS-cohort distances
}
