% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mesa_dataset.R
\name{generate_mesa_dataset}
\alias{generate_mesa_dataset}
\title{generates stap dataset using MESA HFS-cohort distances}
\usage{
generate_mesa_dataset(seed = NULL, num_subj = 100, max_dist = 10,
  alpha = 23, delta = -2.2, beta = 0.1, beta_bar = 0.5,
  theta = 0.7, sigma = 1, W = function(x) exp(-x))
}
\arguments{
\item{seed}{an integer for set.seed()}

\item{num_subj}{number of subjects to simulate}

\item{max_dist}{upper bound on distances to include}

\item{alpha}{true intercept value}

\item{delta}{confounder effect}

\item{beta}{true bef effect}

\item{theta}{true spatial scale}

\item{W}{weight function}
}
\description{
generates stap dataset using MESA HFS-cohort distances
}
